requirements.txt
pandas 
dotenv
google.generativeai
pyarrow
streamlit